requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2
